OneQubitTest, DiffEqTest, and serial OneQubitStochasticTest functional
